moved several functions from netflow to utils where they should have been
